chore(refactor): extract rowCountUtils from modelUtils
